requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
cloud-sql-python-connector==1.10.0
colorlog==6.9.0
cryptography==42.0.8
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Caching==2.3.1
Flask-Cors==4.0.1
Flask-Limiter==3.12
frozenlist==1.4.1
GeoAlchemy2==0.15.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-secret-manager==2.20.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
iniconfig==2.0.0
ipython==8.12.3
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
limits==5.0.0
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1706899926732/work
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
msgpack==1.0.8
multidict==6.0.5
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.0.0
ordered-set==4.1.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pg8000==1.31.2
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.1.2
pyproj==3.6.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
scramp==1.4.5
shapely @ file:///home/conda/feedstock_root/build_artifacts/shapely_1743678311120/work
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.31
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wrapt==1.17.2
yarg==0.1.9
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
